rbcltactor: Mark the animation object
A mark method is added to mark the animation object of an actor. We
need to mark the animation object because this is accessible from Ruby
so it may have a Ruby proxy object attached. The animation object is
not a property so it wouldn't get marked any other way.